pyproject.toml
[project]
name = "tutor-api-gdg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.11",
    "firebase-admin>=6.7.0",
    "google>=3.0.0",
    "google-genai>=1.7.0",
    "google-generativeai>=0.8.4",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]
